[AV2-842] moved apollo to separate module
Purpose: make use of endpoints from mutable domain
