+++ src/app/shared/request/request.service.ts
@@ -7,11 +7,8 @@ import { UtilsService } from '@services/utils.service';
 import { BrowserStorageService } from '@services/storage.service';
 import { environment } from '@environments/environment';
 import { NewRelicService } from '@shared/new-relic/new-relic.service';
-import { Apollo } from 'apollo-angular';
-import { HttpLink } from 'apollo-angular-link-http';
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import gql from 'graphql-tag';
 import { urlFormatter } from 'helper';
+import { ApolloService } from '@shared/apollo/apollo.service';
 
 @Injectable({ providedIn: 'root' })
 export class DevModeService {
@@ -61,7 +58,7 @@ export class RequestService {
     @Optional() config: RequestConfig,
     private newrelic: NewRelicService,
     private devMode: DevModeService,
-    private apollo: Apollo
+    private apolloService: ApolloService
   ) {
     if (config) {
       this.appkey = config.appkey;
@@ -195,11 +192,7 @@ export class RequestService {
    */
   graphQLQuery(query: string, variables?: any, options?: any): Observable<any> {
     options = {...{ noCache: false }, ...options};
-    const watch = this.apollo.watchQuery({
-      query: gql(query),
-      variables: variables || {},
-      fetchPolicy: options.noCache ? 'no-cache' : 'cache-and-network'
-    });
+    const watch = this.apolloService.graphQLQuery(query, variables, options);
     return watch.valueChanges
       .pipe(map(response => {
         this._refreshApikey(response);
@@ -214,10 +207,7 @@ export class RequestService {
    *
    */
   graphQLMutate(query: string, variables = {}): Observable<any> {
-    return this.apollo.mutate({
-      mutation: gql(query),
-      variables: variables
-    })
+    return this.apolloService.graphQLMutate(query, variables)
       .pipe(
         concatMap(response => {
           this._refreshApikey(response);
@@ -233,11 +223,7 @@ export class RequestService {
    */
   chatGraphQLQuery(query: string, variables?: any, options?: any): Observable<any> {
     options = {...{ noCache: false }, ...options};
-    const watch = this.apollo.use('chat').watchQuery({
-      query: gql(query),
-      variables: variables || {},
-      fetchPolicy: options.noCache ? 'no-cache' : 'cache-and-network'
-    });
+    const watch = this.apolloService.chatGraphQLQuery(query, variables, options);
     return watch.valueChanges
       .pipe(map(response => {
         this._refreshApikey(response);
@@ -249,11 +235,7 @@ export class RequestService {
   }
 
   chatGraphQLMutate(query: string, variables = {}): Observable<any> {
-    return this.apollo.use('chat').mutate({
-      mutation: gql(query),
-      variables: variables
-    })
-      .pipe(
+    return this.apolloService.chatGraphQLMutate(query, variables).pipe(
         concatMap(response => {
           // this._refreshApikey(response);
           return of(response);
@@ -297,7 +279,7 @@ export class RequestService {
    *
    * @returns {string}
    */
-  public getAppkey() {
+  public getAppkey(): string {
     return this.appkey;
   }
